* New files doxygen.sh and .cfg to generate doxygen documentation. * New machinery for hard diffraction now in place. Still to be   further checked, so do not try it out yet. * New switch POWHEG:QEDveto has been introduced to steer the treatment   of non-QCD radiation in the POWHEG implementation. * New option in the HepMC interface, whereby PYTHIA particles can be   appended to an existing HepMC event. (Can still be changed if desired.) * Updated initialization of LHEF files to ignore file contents enclosed   in comment tags. * Inserted missing endtag that corrupted the Tunes.php page. * Minor fix for polarization sign when the tau polarization is forced. * Some minor bug fixes in the SUSY code, and some speed optimization. * Clarification on impact-parameter-enhancement calculation.
(Steve: mass mMin still on to-do list.)



git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/pythia8/pythia82/trunk@361 ffdf98f5-c1f1-482d-8be8-39a3f79b3492

+++ include/Pythia8/SusyLesHouches.h
@@ -640,7 +640,7 @@ public:
 
   // Access/change verbose setting
   int verbose() {return verboseSav;}
-  void verbose(double verboseIn) {verboseSav = verboseIn;}
+  void verbose(int verboseIn) {verboseSav = verboseIn;}
 
   // Output of messages from SLHA interface
   void message(int, string,string ,int line=0);
